Fixing the hierarchy of field handlers
* Introduced BaseSimpleFieldHandler class
* Adjusted simple field handlers to extend BaseSimpleFieldHandler
* Introduced BaseTimeFieldHandler class
* Adjusted date/time field handlers to extend BaseTimeFieldHandler
