+++ src/FieldHandlers/BaseSimpleFieldHandler.php
@@ -1,0 +1,14 @@
+<?php
+namespace CsvMigrations\FieldHandlers;
+
+use CsvMigrations\FieldHandlers\BaseFieldHandler;
+
+/**
+ * BaseSimpleFieldHandler
+ *
+ * This class provides the fallback functionality that
+ * is common to simple field handlers.
+ */
+abstract class BaseSimpleFieldHandler extends BaseFieldHandler
+{
+}
