+++ src/FieldHandlers/DatetimeFieldHandler.php
@@ -2,9 +2,9 @@
 namespace CsvMigrations\FieldHandlers;
 
 use Cake\I18n\Time;
-use CsvMigrations\FieldHandlers\BaseFieldHandler;
+use CsvMigrations\FieldHandlers\BaseTimeFieldHandler;
 
-class DatetimeFieldHandler extends BaseFieldHandler
+class DatetimeFieldHandler extends BaseTimeFieldHandler
 {
     /**
      * Database field type
@@ -137,28 +137,6 @@ class DatetimeFieldHandler extends BaseFieldHandler
                     'block' => 'css'
                 ]
             ]
-        ];
-    }
-
-    public function getSearchOperators()
-    {
-        return [
-            'is' => [
-                'label' => 'is',
-                'operator' => 'IN',
-            ],
-            'is_not' => [
-                'label' => 'is not',
-                'operator' => 'NOT IN',
-            ],
-            'greater' => [
-                'label' => 'from',
-                'operator' => '>',
-            ],
-            'less' => [
-                'label' => 'to',
-                'operator' => '<',
-            ],
         ];
     }
 }
