+++ src/FieldHandlers/EmailFieldHandler.php
@@ -1,9 +1,9 @@
 <?php
 namespace CsvMigrations\FieldHandlers;
 
-use CsvMigrations\FieldHandlers\BaseFieldHandler;
+use CsvMigrations\FieldHandlers\BaseSimpleFieldHandler;
 
-class EmailFieldHandler extends BaseFieldHandler
+class EmailFieldHandler extends BaseSimpleFieldHandler
 {
     /**
      * HTML form field type
