+++ src/FieldHandlers/ReminderFieldHandler.php
@@ -5,5 +5,4 @@ use CsvMigrations\FieldHandlers\DatetimeFieldHandler;
 
 class ReminderFieldHandler extends DatetimeFieldHandler
 {
-    //and some magic goes here
 }
