+++ src/FieldHandlers/UuidFieldHandler.php
@@ -1,9 +1,9 @@
 <?php
 namespace CsvMigrations\FieldHandlers;
 
-use CsvMigrations\FieldHandlers\BaseFieldHandler;
+use CsvMigrations\FieldHandlers\BaseSimpleFieldHandler;
 
-class UuidFieldHandler extends BaseFieldHandler
+class UuidFieldHandler extends BaseSimpleFieldHandler
 {
     /**
      * Field type
